error_handle() accepts error and solution and returns dict for tablulate response

+++ jobbergate_cli/jobbergate_api_wrapper.py
@@ -85,7 +85,7 @@ class JobbergateApi:
             response = requests.delete(
                 endpoint,
                 headers={'Authorization': 'JWT ' + self.token},
-                verify=False).text
+                verify=False)
 
         if method == "POST":
             try:
@@ -170,6 +170,14 @@ class JobbergateApi:
 
         return question_list
 
+    def error_handle(self, error, solution):
+        response = {
+            "error": error,
+            "solution": solution
+        }
+        return response
+
+
     # Job Scripts
     @tabulate_decorator
     def list_job_scripts(self, all):
@@ -198,6 +206,21 @@ class JobbergateApi:
                           application_id,
                           param_file,
                           debug):
+
+        if application_id is None:
+            response = self.error_handle(
+                error="--application-id for the job script not defined",
+                solution="Please try again with --application-id specified"
+            )
+            return response
+
+        if job_script_name is None:
+            response = self.error_handle(
+                error="--name for the job script not defined",
+                solution="Please try again with --name specified"
+            )
+            return response
+
         data = self.job_script_config
         data['job_script_name'] = job_script_name
         data['application'] = application_id
@@ -206,10 +229,10 @@ class JobbergateApi:
         if param_file:
             is_param_file = os.path.isfile(param_file)
             if is_param_file is False:
-                response = {
-                    "error": f"invalid --parameter-file supplied, could not find: {param_file}",
-                    "solution": f"Please provide the full path to a valid parameter file"
-                }
+                response = self.error_handle(
+                    error=f"invalid --parameter-file supplied, could not find: {param_file}",
+                    solution="Please provide the full path to a valid parameter file"
+                )
                 return response
 
             files = {'upload_file': open(param_file, 'rb')}
@@ -318,6 +341,13 @@ class JobbergateApi:
     def get_job_script(self,
                        job_script_id,
                        as_str):
+        if job_script_id is None:
+            response = self.error_handle(
+                error="--id not define",
+                solution="Please try again with --id specified"
+            )
+            return response
+
         response = self.jobbergate_request(
             method="GET",
             endpoint=f"{self.api_endpoint}/job-script/{job_script_id}"
@@ -339,6 +369,13 @@ class JobbergateApi:
     @tabulate_decorator
     def update_job_script(self,
                           job_script_id):
+        if job_script_id is None:
+            response = self.error_handle(
+                error="--id not defined",
+                solution="Please try again with --id specified"
+            )
+            return response
+
         data = self.jobbergate_request(
             method="GET",
             endpoint=f"{self.api_endpoint}/job-script/{job_script_id}"
@@ -352,8 +389,16 @@ class JobbergateApi:
 
         return response
 
+    @tabulate_decorator
     def delete_job_script(self,
                           job_script_id):
+        if job_script_id is None:
+            response = self.error_handle(
+                error="--id not defined",
+                solution="Please try again with --job-script-id specified"
+            )
+            return response
+
         response = self.jobbergate_request(
             method="DELETE",
             endpoint=f"{self.api_endpoint}/job-script/{job_script_id}"
@@ -384,9 +429,16 @@ class JobbergateApi:
 
     @tabulate_decorator
     def create_job_submission(self,
-                              job_submission_name,
                               job_script_id,
-                              render_only):
+                              render_only,
+                              job_submission_name=""):
+        if job_script_id is None:
+            response = self.error_handle(
+                error="--job-script-id not defined",
+                solution="Please try again with --job-script-id specified"
+            )
+            return response
+
         data = self.job_submission_config
         data['job_submission_name'] = job_submission_name
         data['job_script'] = job_script_id
@@ -437,6 +489,12 @@ class JobbergateApi:
     @tabulate_decorator
     def get_job_submission(self,
                            job_submission_id):
+        if job_submission_id is None:
+            response = self.error_handle(
+                error="--id not defined",
+                solution="Please try again with --id specified"
+            )
+            return response
         try:
             response = self.jobbergate_request(
                 method="GET",
@@ -450,6 +508,13 @@ class JobbergateApi:
     @tabulate_decorator
     def update_job_submission(self,
                               job_submission_id):
+        if job_submission_id is None:
+            response = self.error_handle(
+                error="--id not defined",
+                solution="Please try again with --id specified"
+            )
+            return response
+
         data = self.jobbergate_request(
             method="GET",
             endpoint=f"{self.api_endpoint}/job-submission/{job_submission_id}"
@@ -462,15 +527,31 @@ class JobbergateApi:
         )
         return response
 
+    @tabulate_decorator
     def delete_job_submission(self,
                               job_submission_id):
-        try:
-            response = self.jobbergate_request(
-                method="DELETE",
-                endpoint=f"{self.api_endpoint}/job-submission/{job_submission_id}"
+        if job_submission_id is None:
+            response = self.error_handle(
+                error="--id not defined",
+                solution="Please try again with --id specified"
             )
-        except:
-            response = f"Failed to DELETE job submission id: {job_submission_id}"
+            return response
+
+
+        response = self.jobbergate_request(
+            method="DELETE",
+            endpoint=f"{self.api_endpoint}/job-submission/{job_submission_id}"
+        )
+        if response.status_code == 404:
+            response = self.error_handle(
+                error="",
+                solution=""
+            )
+            response = {
+                "error": f"Failed to DELETE job submission id: {job_submission_id}",
+                "solution": f"Please try again with a valid job submission id"
+            }
+            return response
         return response
 
     # Applications
@@ -505,34 +586,48 @@ class JobbergateApi:
         create an application based on path provided by the user
         '''
 
+        parameter_check = []
+        if application_name is None:
+            response = self.error_handle(
+                error="--name not defined",
+                solution="Please try again with --name specified for the application"
+            )
+            parameter_check.append(response)
+
+        if application_path is None:
+            response = self.error_handle(
+                error="--application-path not defined",
+                solution="Please try again with --application-path specified for the application"
+            )
+            parameter_check.append(response)
+        if len(parameter_check) > 0:
+            response = parameter_check
+            return response
+
         #check for required files
+        error_check = []
         is_dir = os.path.isdir(application_path)
         is_app_file = os.path.isfile(f"{application_path}/{APPLICATION_FILENAME}")
         is_config_file = os.path.isfile(f"{application_path}/{CONFIG_FILENAME}")
 
-        error_check = []
         if is_dir is False:
-            error_check.append(
-                {
-                    "error": "invalid application path supplied",
-                    "solution": f"{application_path} is invalid, please review and try again"
-
-                }
+            check = self.error_handle(
+                error="invalid application path supplied",
+                solution=f"{application_path} is invalid, please review and try again"
             )
+            error_check.append(check)
         if is_app_file is False:
-            error_check.append(
-                {
-                    "error": f"Could not find {APPLICATION_FILENAME} in {application_path}",
-                    "solution": f"Please ensure {APPLICATION_FILENAME} is in application path provided"
-                }
+            check = self.error_handle(
+                error=f"Could not find {APPLICATION_FILENAME} in {application_path}",
+                solution=f"Please ensure {APPLICATION_FILENAME} is in application path provided"
             )
+            error_check.append(check)
         if is_config_file is False:
-            error_check.append(
-                {
-                    "error": f"Could not find {CONFIG_FILENAME} in {application_path}",
-                    "solution": f"Please ensure {CONFIG_FILENAME} is in application path provided"
-                }
+            check = self.error_handle(
+                error=f"Could not find {CONFIG_FILENAME} in {application_path}",
+                solution=f"Please ensure {CONFIG_FILENAME} is in application path provided"
             )
+            error_check.append(check)
 
         if len(error_check) > 0:
             response = error_check
@@ -602,14 +697,18 @@ class JobbergateApi:
         )
         return response
 
+    @tabulate_decorator
     def delete_application(self,
                            application_id):
-        try:
-            response = self.jobbergate_request(
-                method="DELETE",
-                endpoint=f"{self.api_endpoint}/application/{application_id}"
+        response = self.jobbergate_request(
+            method="DELETE",
+            endpoint=f"{self.api_endpoint}/application/{application_id}"
+        )
+        if response.status_code == 404:
+            response = self.error_handle(
+                error=f"Failed to DELETE job submission id: {application_id}",
+                solution="Please try again with a valid job submission id"
             )
-        except:
-            response = f"Failed to DELETE application id: {application_id}"
+            return response
 
         return response
